do some mored cleanup
