--- components/ExpanderDiv.tsx
@@ -1,8 +1,0 @@
-import styled from "styled-components";
-
-export const ExpanderDiv = styled.div`
-  @media (min-width: 1024px) {
-    margin-left: -80px;
-    margin-right: -80px;
-  }
-`;
