+++ components/Footer.tsx
@@ -1,108 +1,39 @@
-import { Twitter, GitHub } from "react-feather";
-import { Flex } from "rebass/styled-components";
-import { ReactNode, ElementType } from "react";
-import Text from "./Text";
-import Link from "./Link";
-import Layout from "./Layout";
-import Icon from "./Icon";
-import { TwoColumnGrid } from "./TwoColumnGrid";
-import { textColor } from "./textColor";
+import { Twitter, Github } from "lucide-react";
+import { Separator } from "./ui/separator";
+import { Link } from "./ui/link";
 
-interface FooterColumnProps {
-  title: string;
-  width?: number;
-  last?: boolean;
-  children: ReactNode;
-}
-
-interface FooterItemProps {
-  IconComp?: ElementType;
-  children: ReactNode;
-  href: string;
-  rel?: string;
-  prefetch?: boolean;
-}
-const FooterColumn = (props: FooterColumnProps) => {
+export function Footer() {
   return (
-    <Flex
-      flexDirection="column"
-      flex={1}
-      mb={4}
-      mr={props?.last === true ? 0 : 4}
-    >
-      <Text mb={3} fontWeight="bold">
-        {props.title}
-      </Text>
-      {props.children}
-    </Flex>
-  );
-};
-
-const FooterItem = ({
-  IconComp,
-  href,
-  rel,
-  children,
-  prefetch,
-}: FooterItemProps) => {
-  return (
-    <Text as="div" my={[3, 1]}>
-      <Link
-        href={href}
-        rel={rel}
-        prefetch={prefetch}
-        underline
-        external
-        css={textColor}
-      >
-        {IconComp && (
-          <Icon mr={2} ml={0}>
-            <IconComp size="1em" />
-          </Icon>
-        )}
-        {children}
-      </Link>
-    </Text>
-  );
-};
-
-const Footer = (): JSX.Element => {
-  return (
-    <Flex
-      py={5}
-      mt={5}
-      as="footer"
-      bg="var(--color-secondary-background)"
-      px={[3, 4]}
-    >
-      <Layout width={[1, 0.5]}>
-        <TwoColumnGrid>
-          <FooterColumn title="About this website">
-            <Text lineHeight={1.5}>
+    <>
+      <div className="flex flex-col gap-4 bg-secondary-background">
+        <footer className="p-8 grid sm:grid-cols-2 gap-16 place-content-center">
+          <div className="flex flex-1 flex-col gap-4 max-w-80 place-self-end">
+            <h3 className="font-bold text-xl">About this website</h3>
+            <p className="font-medium">
               I&apos;m Hiranmaya Gundu, a dev exploring web and systems
               programming. This is my personal blog!
-            </Text>
-          </FooterColumn>
-          <FooterColumn last title="Social Media">
-            <FooterItem
-              href="https://github.com/HiranmayaGundu/hmgundu.dev"
-              IconComp={GitHub}
-              rel="me noopener"
-            >
-              View the source on GitHub
-            </FooterItem>
-            <FooterItem
-              href="https://twitter.com/hiranmayagundu"
-              IconComp={Twitter}
-              rel="me noopener"
-            >
-              You can follow me on twitter
-            </FooterItem>
-          </FooterColumn>
-        </TwoColumnGrid>
-      </Layout>
-    </Flex>
+            </p>
+          </div>
+          <div className="flex flex-1 flex-col gap-4 max-w-80 place-self-start">
+            <h3 className="font-bold text-xl">Social Media</h3>
+            <div className="flex gap-2">
+              <Github />
+              <Link
+                href="https://github.com/HiranmayaGundu/hmgundu.dev"
+                rel="me noopener"
+              >
+                View the source on GitHub
+              </Link>
+            </div>
+            <div className="flex gap-2">
+              <Twitter />
+              <Link href="https://twitter.com/hiranmayagundu" rel="me noopener">
+                Follow me on twitter
+              </Link>
+            </div>
+          </div>
+        </footer>
+      </div>
+    </>
   );
-};
-
-export default Footer;
+}
