--- components/Layout.tsx
@@ -1,9 +1,0 @@
-import { Box, BoxProps } from "rebass/styled-components";
-import styled from "styled-components";
-
-const Layout: React.FC<BoxProps> = styled(Box)`
-  max-width: 800px;
-  margin: 0 auto;
-`;
-
-export default Layout;
