--- components/textColor.tsx
@@ -1,5 +1,0 @@
-import { css } from "styled-components";
-
-export const textColor = css`
-  color: var(--color-text);
-`;
