+++ package.json
@@ -23,7 +23,6 @@
     "@tippyjs/react": "^4.2.5",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.0",
-    "conditional-wrap": "^1.0.0",
     "date-fns": "^2.20.1",
     "fast-glob": "^3.3.2",
     "lodash.kebabcase": "^4.1.1",
@@ -72,7 +71,6 @@
     "eslint-plugin-react-hooks": "^4.4.0",
     "husky": "^7.0.4",
     "install": "^0.13.0",
-    "jest": "^27.5.1",
     "lint-staged": ">=12.3.7",
     "postcss": "^8.4.35",
     "prettier": "^2.6.2",
